+++ src/api/review/api.ts
@@ -96,31 +96,6 @@ export const updateReviewApi = async (
   return res.data;
 };
 
-// like review
-export const likeReviewApi = async (reviewId: number, accessToken: string) => {
-  const config = {
-    headers: { Authorization: `Bearer ${accessToken}` },
-  };
-  const res = await axios.post(
-    `/api/v1/reviews/${reviewId}/likes`,
-    null,
-    config
-  );
-
-  return res.data;
-};
-
-export const dislikeReviewApi = async (
-  reviewId: number,
-  accessToken: string
-) => {
-  const config = {
-    headers: { Authorization: `Bearer ${accessToken}` },
-  };
-  const res = await axios.delete(`/api/v1/reviews/${reviewId}/likes`, config);
-  return res.data;
-};
-
 // get one review by ID
 export const getSingleReviewApi = async (reviewId?: number | null) => {
   if (reviewId === null || reviewId === undefined) {
